refactor: create FoodList component to display food products
This commit adds a new component called FoodList, which is responsible for rendering a list of food products. The component imports the ProductCard component from "@/components/product-card" and uses the foodList constant from "@/constants" to populate the list. The list is displayed in a grid layout with 1 column on small screens, 2 columns on medium screens, and 3 columns on large screens. Each product is rendered using the ProductCard component, passing the product information and specifying the type as "food". This component is used to showcase the available food products on the website.